update tests to reflect chunking being moved out of default entrypoint
